requirements.txt
uvicorn==0.22.0
pydantic==1.10.8
pymongo==4.3.3
fastapi>=0.65.0
streamlit
numpy
sympy
langchain
langchain_community
langchain_openai
matplotlib
python-dotenv
bs4
faiss-cpu
beautifulsoup4 
wordcloud
requests

httpx
asyncio